% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gain_for_solutions.R
\name{calculate_gain_for_solutions}
\alias{calculate_gain_for_solutions}
\title{Calculate gain for solutions}
\usage{
calculate_gain_for_solutions(clusterings, contextual_variable)
}
\arguments{
\item{clusterings}{List — a list of cluster solutions from the `nTARP`
function output (`"Clusterings"`)}

\item{contextual_variable}{Vector of integers or characters — variable to use as the basis
for comparing clusters}
}
\value{
A list containing:
  (1) Gains for each pair of clusters, and
  (2) Output from the `evaluate_cluster_solution` function containing full
      information on each solution's Gini indices and distribution of groups.
}
\description{
Calculate gain for solutions
}
\keyword{(n_1}
\keyword{(n_2}
\keyword{(parent)}
\keyword{*}
\keyword{+}
\keyword{-}
\keyword{/}
\keyword{=}
\keyword{Calculates}
\keyword{Gain}
\keyword{Gini}
\keyword{Gini(cluster_1)}
\keyword{Gini(cluster_2)}
\keyword{Gini(parent)}
\keyword{The}
\keyword{[}
\keyword{]}
\keyword{`nTARP`}
\keyword{`nTARP`.}
\keyword{a}
\keyword{and}
\keyword{are}
\keyword{as:}
\keyword{based}
\keyword{bisecting}
\keyword{by}
\keyword{cluster}
\keyword{cluster,}
\keyword{clustering}
\keyword{clusters}
\keyword{clusters.}
\keyword{contextual}
\keyword{defined}
\keyword{derived}
\keyword{determine}
\keyword{each}
\keyword{explaining}
\keyword{for}
\keyword{from}
\keyword{gain}
\keyword{gains}
\keyword{helpful}
\keyword{in}
\keyword{index}
\keyword{internal}
\keyword{is}
\keyword{it}
\keyword{most}
\keyword{n}
\keyword{n)}
\keyword{n_1}
\keyword{n_2}
\keyword{obtained}
\keyword{of}
\keyword{on}
\keyword{order}
\keyword{original}
\keyword{output}
\keyword{resulting}
\keyword{size}
\keyword{sizes}
\keyword{solution}
\keyword{split}
\keyword{structure.}
\keyword{the}
\keyword{to}
\keyword{two}
\keyword{user-specified}
\keyword{using}
\keyword{variable,}
\keyword{where}
\keyword{which}
