% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids_from_clusterings.R
\name{get_ids_from_clusterings}
\alias{get_ids_from_clusterings}
\title{Get IDs from clusterings}
\usage{
get_ids_from_clusterings(ids, cluster_member_vector)
}
\arguments{
\item{ids}{Numeric or character vector — identifying labels for individuals in the clusters}

\item{cluster_member_vector}{Numeric vector — cluster assignments from `nTARP`}
}
\value{
A list with two entries:
  - One containing the IDs in cluster 1
  - One containing the IDs in cluster 2
}
\description{
#' @keywords  internal
}
\details{
Pulls the IDs used to identify observations in cluster solutions so they
can be used for other purposes, including tracking observations between solutions.
}
