% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_best_solution_and_gain.R
\name{pull_best_solution_and_gain}
\alias{pull_best_solution_and_gain}
\title{Get IDs from clusterings}
\usage{
pull_best_solution_and_gain(
  gains_list,
  full_gains_list,
  clusterings,
  ids,
  contextual_variable
)
}
\arguments{
\item{gains_list}{Numeric vector — gain values from the
`calculate_gains_for_solution` output (`Gains`)}

\item{full_gains_list}{List — gain values from the
`calculate_gains_for_solution` output (`FullGainInformation`)}

\item{clusterings}{List — a list of cluster solutions from the `nTARP`
function output (`Clusterings`)}

\item{ids}{Numeric or character vector — identifying labels for individuals in the clusters}

\item{contextual_variable}{Vector of integers or characters — variable used as the basis
for comparing clusters}
}
\value{
A list with two entries:
  - One containing the IDs in cluster 1
  - One containing the IDs in cluster 2
}
\description{
#' @keywords  internal
}
\details{
Pulls the IDs used to identify observations in cluster solutions so that
they can be used for other purposes, including tracking observations
between solutions.
}
