% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_clusters.R
\name{consolidate_clusters}
\alias{consolidate_clusters}
\title{Merge clusters as post-processing}
\usage{
consolidate_clusters(
  cluster_path_matrix,
  first_cluster_to_combine,
  second_cluster_to_combine
)
}
\arguments{
\item{cluster_path_matrix}{Data frame — output of
`build_solution_from_labeled_clusters` showing which branch each
observation belongs to from the `nTARP` clustering}

\item{first_cluster_to_combine}{Numeric — label of the first cluster to merge}

\item{second_cluster_to_combine}{Numeric — label of the second cluster to merge}
}
\value{
A data frame in the same format as the first argument, with the
  final column showing the cluster IDs relabeled based on the chosen merge.
}
\description{
This function combines user-specified clusters after `nTARP` has run.
The input to the function is the output of the
`build_solution_from_labeled_clusters` function, along with the two
cluster labels the user would like to merge. The output is returned
in the same format as `build_solution_from_labeled_clusters`, with
the final solution labels renamed to reflect the merging.
}
